refactor: 問題一覧の表示に関する処理を、コンポーネントに分割
